Revert "small changes"
This reverts commit c3b8efa1f8bae1466ab6fe3153f069297b5c5ab7.

+++ Assets/Scripts/CreatureManagerScript.cs
@@ -8,7 +8,7 @@ public class CreatureManagerScript : MonoBehaviour
 	[SerializeField] private GameManager gameManager;
 
 	[SerializeField, Tooltip("second * 60fps")] private int[] spawnIntervalTimeArray;
-	[SerializeField, Tooltip("creatureSpeedUp")] private float speedCreatureUp;
+	[SerializeField, Tooltip("creatureSpeedUp")] private int speedCreatureUp;
 	[SerializeField, Tooltip("second * 60fps")] private int speedUpIntervalTime;
 	private int speedUpLevel;
 	private int frame;
